% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bauges_IFN_Stands}
\alias{Bauges_IFN_Stands}
\title{Bauges_IFN_Stands dataset.}
\format{
An object of class \code{data.frame} with 234 rows and 15 columns.
}
\usage{
Bauges_IFN_Stands
}
\description{
Bauges dataset from IFN database.
Stand characteristics.
}
\keyword{datasets}
