% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bauges_Virtual}
\alias{Bauges_Virtual}
\title{Bauges_Virtual dataset.}
\format{
A data frame with seven variables: \code{ID}, \code{Area},
\code{Dg}, \code{BA} \code{Gini} \code{NHA} \code{Compo}
}
\usage{
Bauges_Virtual
}
\description{
Bauges dataset from Arnaud's virtual choice of stands.
}
\keyword{datasets}
