% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Boxplots.R
\name{Plot_Boxplots}
\alias{Plot_Boxplots}
\title{Plot boxplots}
\usage{
Plot_Boxplots(df, quantiVars, factorCol = NULL, factorGrid = NULL)
}
\arguments{
\item{df}{dataframe in which data to plot are stored}

\item{quantiVars}{strings vector of quantitative variables to plot}

\item{factorCol}{string name of the factor variable to distinct with colors}

\item{factorGrid}{string name of the factor variable to distinct within a plot grid}
}
\value{
ggplot2 grid of boxplots for each variable.
}
\description{
Plot the boxplots for all given quantitative variables, according given factors, of the given dataset.
}
\examples{
Plot_Boxplots(data, varNames, factors[1])
Plot_Boxplots(data, Yquanti, NULL, factors[2])
Plot_Boxplots(data, Yquanti, factors[1], factors[2])
}
