% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Distributions.R
\name{Plot_Distributions}
\alias{Plot_Distributions}
\title{Plot the distributions}
\usage{
Plot_Distributions(df, quantiVars)
}
\arguments{
\item{df}{dataframe in which data to plot are stored}

\item{quantiVars}{strings vector of quantitative variables to plot}
}
\value{
ggplot2 histogram for each variable
}
\description{
Plot the distribution histograms for all given quantitative variables of the given dataset.
}
\examples{
Plot_Distributions(data, varNames)
}
