% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Homoscedasticity.R
\name{Plot_Homoscedasticity}
\alias{Plot_Homoscedasticity}
\title{Plot for analysing homoscedasticity}
\usage{
Plot_Homoscedasticity(model)
}
\arguments{
\item{model}{linear model to check homoscedasticity assumption}
}
\value{
ggplot2 fitted VS residuals scatterplot
}
\description{
Plot a scatter plot fitted values VS residuals of a linear model
}
\examples{
model <- lm(y~x)
Plot_Homoscedasticity(model)
}
