% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Normalities.R
\name{Plot_Normalities}
\alias{Plot_Normalities}
\title{Plot for analysing normality}
\usage{
Plot_Normalities(df, quantiVars)
}
\arguments{
\item{df}{dataframe in which data to plot are stored}

\item{quantiVars}{strings vector of quantitative variables to plot}
}
\value{
ggplot2 graphics with QQ-plot and shapiro test p-value.
}
\description{
Plot a QQ-plot and perform a Shapiro test for given variables in a given dataset.
}
\examples{
Plot_Normalities(data, varNames)
}
