% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Outliers.R
\name{Plot_Outliers}
\alias{Plot_Outliers}
\title{Plot graphics for detecting outliers}
\usage{
Plot_Outliers(df, quantiVars)
}
\arguments{
\item{df}{dataframe in which data to plot are stored}

\item{quantiVars}{strings vector of quantitative variables to plot}
}
\value{
ggplot2 graphics with boxplot and cleveland plot in a grid for each variable.
}
\description{
Plot a boxplot and cleveland plot for all given quantitative variables of the given dataset.
It aims to detect outliers within a given distribution.
}
\examples{
Plot_Outliers(data, varNames)
}
