% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Relationships.R
\name{Plot_Relationships}
\alias{Plot_Relationships}
\title{Plot relationships between variables}
\usage{
Plot_Relationships(
  df,
  XquantiVars,
  YquantiVars,
  factorCol = NULL,
  factorGrid = NULL
)
}
\arguments{
\item{df}{dataframe in which data to plot are stored}

\item{XquantiVars}{strings vector of quantitative explanatory variables to plot}

\item{YquantiVars}{strings vector of quantitative response variables to plot}

\item{factorCol}{string name of the factor variable to distinct with colors}

\item{factorGrid}{string name of the factor variable to distinct within a plot grid}
}
\value{
ggplot2 grid of boxplots for each variable.
}
\description{
Plot the relationships between given X (explanatory) and Y (response) variables.
According to given factors of a given dataset.
}
\examples{
Plot_Relationships(data, XvarNames, XvarNames)
Plot_Relationships(data, XvarNames, XvarNames, NULL, Xfactors[2])
Plot_Relationships(data, XvarNames, YvarNames)
Plot_Relationships(data, XvarNames, YvarNames, Xfactors[1], Xfactors[2])
}
