% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bauges_IFN_Trees}
\alias{Bauges_IFN_Trees}
\title{Bauges_Virtual.}
\format{
An object of class \code{data.frame} with 3697 rows and 1 columns.
}
\usage{
Bauges_IFN_Trees
}
\description{
Bauges dataset from IFN database.
List of trees in a specified stand.
}
\keyword{datasets}
